# frontend/src/lib/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RegPulse API",
    "description": "RBI Regulatory Intelligence Platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/debates": {
      "get": {
        "tags": [
          "debates"
        ],
        "summary": "List Debates",
        "description": "List debates (currently always empty until slice 9).",
        "operationId": "list_debates_api_v1_debates_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/DebateStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DebateListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "debates"
        ],
        "summary": "Create Debate",
        "operationId": "create_debate_api_v1_debates_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DebateCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DebateDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/debates/{debate_id}": {
      "get": {
        "tags": [
          "debates"
        ],
        "summary": "Get Debate",
        "operationId": "get_debate_api_v1_debates__debate_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "debate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Debate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DebateDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/debates/{debate_id}/reply": {
      "post": {
        "tags": [
          "debates"
        ],
        "summary": "Reply To Debate",
        "operationId": "reply_to_debate_api_v1_debates__debate_id__reply_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "debate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Debate Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DebateReplyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DebateDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/debates/{debate_id}/resolve": {
      "patch": {
        "tags": [
          "debates"
        ],
        "summary": "Resolve Debate",
        "operationId": "resolve_debate_api_v1_debates__debate_id__resolve_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "debate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Debate Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DebateResolveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DebateDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/annotations/{annotation_id}": {
      "patch": {
        "tags": [
          "annotations"
        ],
        "summary": "Update Annotation",
        "operationId": "update_annotation_api_v1_annotations__annotation_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "annotation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Annotation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnotationUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnotationDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "annotations"
        ],
        "summary": "Delete Annotation",
        "operationId": "delete_annotation_api_v1_annotations__annotation_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "annotation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Annotation Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Annotation": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "question_id": {
            "type": "string",
            "format": "uuid",
            "title": "Question Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "user_initials": {
            "type": "string",
            "title": "User Initials"
          },
          "text_selection": {
            "type": "string",
            "title": "Text Selection"
          },
          "note": {
            "type": "string",
            "title": "Note"
          },
          "anchor_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anchor Offset"
          },
          "resolved": {
            "type": "boolean",
            "title": "Resolved",
            "default": false
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "question_id",
          "user_id",
          "user_initials",
          "text_selection",
          "note",
          "created_at"
        ],
        "title": "Annotation"
      },
...
      "AnnotationDetailResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/Annotation"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "AnnotationDetailResponse"
      },
...
      "AnnotationUpdateRequest": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "resolved": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved"
          }
        },
        "type": "object",
        "title": "AnnotationUpdateRequest"
      },
...
      "DebateCreateRequest": {
        "properties": {
          "question_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Id"
          },
          "source_circular_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Circular Id"
          },
          "title": {
            "type": "string",
            "maxLength": 240,
            "minLength": 8,
            "title": "Title"
          },
          "opening_text": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 4,
            "title": "Opening Text"
          },
          "initial_stance": {
            "$ref": "#/components/schemas/DebateStance",
            "default": "NEUTRAL"
          }
        },
        "type": "object",
        "required": [
          "title",
          "opening_text"
        ],
        "title": "DebateCreateRequest"
      },
      "DebateDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "user_initials": {
            "type": "string",
            "title": "User Initials"
          },
          "question_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Id"
          },
          "source_circular_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Circular Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "$ref": "#/components/schemas/DebateStatus"
          },
          "agree_count": {
            "type": "integer",
            "title": "Agree Count",
            "default": 0
          },
          "disagree_count": {
            "type": "integer",
            "title": "Disagree Count",
            "default": 0
          },
          "reply_count": {
            "type": "integer",
            "title": "Reply Count",
            "default": 0
          },
          "resolution_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Text"
          },
          "resolved_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved By"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "replies": {
            "items": {
              "$ref": "#/components/schemas/DebateReply"
            },
            "type": "array",
            "title": "Replies"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "user_initials",
          "title",
          "status",
          "created_at"
        ],
        "title": "DebateDetail"
      },
      "DebateDetailResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/DebateDetail"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "DebateDetailResponse"
      },
      "DebateItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "user_initials": {
            "type": "string",
            "title": "User Initials"
          },
          "question_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Id"
          },
          "source_circular_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Circular Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "$ref": "#/components/schemas/DebateStatus"
          },
          "agree_count": {
            "type": "integer",
            "title": "Agree Count",
            "default": 0
          },
          "disagree_count": {
            "type": "integer",
            "title": "Disagree Count",
            "default": 0
          },
          "reply_count": {
            "type": "integer",
            "title": "Reply Count",
            "default": 0
          },
          "resolution_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Text"
          },
          "resolved_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved By"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "user_initials",
          "title",
          "status",
          "created_at"
        ],
        "title": "DebateItem"
      },
      "DebateListResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/DebateItem"
            },
            "type": "array",
            "title": "Data"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          },
          "open_count": {
            "type": "integer",
            "title": "Open Count",
            "default": 0
          }
        },
        "type": "object",
        "title": "DebateListResponse"
      },
      "DebateReply": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "debate_id": {
            "type": "string",
            "format": "uuid",
            "title": "Debate Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "user_initials": {
            "type": "string",
            "title": "User Initials"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "stance": {
            "$ref": "#/components/schemas/DebateStance"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "debate_id",
          "user_id",
          "user_initials",
          "text",
          "stance",
          "created_at"
        ],
        "title": "DebateReply"
      },
      "DebateReplyRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 2,
            "title": "Text"
          },
          "stance": {
            "$ref": "#/components/schemas/DebateStance",
            "default": "NEUTRAL"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "DebateReplyRequest"
      },
      "DebateResolveRequest": {
        "properties": {
          "resolution_text": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 4,
            "title": "Resolution Text"
          }
        },
        "type": "object",
        "required": [
          "resolution_text"
        ],
        "title": "DebateResolveRequest"
      },
      "DebateStance": {
        "type": "string",
        "enum": [
          "AGREE",
          "DISAGREE",
          "NEUTRAL"
        ],
        "title": "DebateStance"
      },
      "DebateStatus": {
        "type": "string",
        "enum": [
          "OPEN",
          "RESOLVED"
        ],
        "title": "DebateStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}